+++ plugins/mesh/snow/object/snow.h
@@ -131,7 +131,7 @@ public:
   // Redirect these functions to csParticleSystem
   virtual void SetParticleCount (int num)
   { csParticleSystem::SetParticleCount (num); }
-  virtual int GetParticleCount () const
+  virtual size_t GetParticleCount () const
   { return csParticleSystem::GetParticleCount (); }
 };
